[STEP09] Filter, Interceptor, Exception

+++ src/main/java/kr/hhplus/be/application/payment/dto/PaymentCommand.java
@@ -2,6 +2,8 @@ package kr.hhplus.be.application.payment.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @AllArgsConstructor
@@ -9,5 +11,5 @@ import lombok.*;
 public class PaymentCommand {
     private Long userId;
     private Long orderId;
-    private Integer payAmount;
+    private BigDecimal payAmount;
 }
