+++ src/main/java/kr/hhplus/be/domain/order/dto/OrderDTO.java
@@ -2,13 +2,15 @@ package kr.hhplus.be.domain.order.dto;
 
 import kr.hhplus.be.application.order.dto.OrderItems;
 
+import java.math.BigDecimal;
+
 public class OrderDTO {
     private Long orderId;
     private Long userId;
     private Long couponPublishId;
-    private int totalAmount;
-    private int discountAmount;
-    private int finalAmount;
+    private BigDecimal totalAmount;
+    private BigDecimal discountAmount;
+    private BigDecimal finalAmount;
     private OrderItems orderItems;
 
 }
